feat(validator): create a validateRequest method that validates the state of the request, and determ
The method starts by removing unwanted characters in the request url, such as query parameters, hash
values, and the dangerous .. characters. Then it checks if the request method is one of GET, HEAD or
OPTIONS. then it runs through the specified public paths and checks if there is any that contains
the request path. It returns the absolute file path if all processes validates, or empty string if
response should be ignored. it also performs check for default documents if request path maps to a
folder in one of the public paths. Note that public paths is the root directory by default unless
otherwise specified. It also performs a check and avoids serving configuration files that starts
with the dot character